Fix util package exports
