+++ packages/utils/src/index.ts
@@ -15,8 +15,8 @@ export { removeUnderscores } from "./removeUnderscores.js";
 export { getShortUniqueDnp } from "./getShortUniqueDnp.js";
 export {
   buildNetworkAlias,
-  getPrivateNetworkAliases,
 } from "./buildNetworkAlias.js";
+export { getPrivateNetworkAliases, } from "./getPrivateNetworkAliases.js";
 export { getIsCore } from "./getIsCore.js";
 export { shell, shellHost, ShellError } from "./shell.js";
 export { normalizeHash } from "./normalizeHash.js";
